+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 <!-- There is always Unreleased section on the top. Subsections (Add, Changed, Fix, Removed) should be Add as needed. -->
 ## Unreleased
 - Update dependencies
+- [**BC**] Remove Graylog
+- [**BC**] Remove ConsoleStyle
+- [**BC**] Remove ApplicationLogger
+- Use common Logger extensions to provide a logger factory
 
 ## 4.2.0 - 2021-03-18
 - Add explicit dependency on `Microsoft.Extensions.Logging (5.0)`